+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## V2.12.0
+### Changed
+- Changes the type of the `additional_info.shipments.receiver_address` attribute from `AdditionalInfoAddress` to `ReceiverAddress`.
+- The `additional_info.shipments.receiver_address` field in this version will no longer accept the attributes: `city`, `country`, `state`, `number` and `complement`.
+
 ## V2.11.2
 ### Fixed
 - Fixes errors pointed out by PHPStan